APP-1249 Create a front end enricher for JIRA comment task
- Fix PR comments
